vim: Move code for delta application and generation to changetracker.lua
