+++ vim-plugin/plugin/ethersync.lua
@@ -1,8 +1,5 @@
 local utils = require("utils")
-local sync = require("vim.lsp.sync")
-
--- Used to note that changes to the buffer should be ignored, and not be sent out as deltas.
-local ignore_edits = false
+local changetracker = require("changetracker")
 
 -- JSON-RPC connection.
 local client
@@ -21,43 +18,6 @@ local daemonRevision = 0
 -- Number of operations we have made.
 local editorRevision = 0
 
--- Used to remember the previous content of the buffer, so that we can
--- calculate the difference between the previous and the current content.
-local prev_lines
-
-local function debug(tbl)
-    if true then
-        client.notify("debug", tbl)
-    end
-end
-
-local function applyDelta(delta)
-    local text_edits = {}
-    for _, replacement in ipairs(delta) do
-        local text_edit = {
-            range = {
-                start = replacement.range.anchor,
-                ["end"] = replacement.range.head,
-            },
-            newText = replacement.replacement,
-        }
-        table.insert(text_edits, text_edit)
-    end
-
-    -- Find correct buffer to apply edits to.
-    local bufnr = vim.uri_to_bufnr("file://" .. theFile)
-
-    ignore_edits = true
-    local changedtick_before = vim.api.nvim_buf_get_changedtick(bufnr)
-    utils.apply_text_edits(text_edits, bufnr, "utf-32")
-    local changedtick_after = vim.api.nvim_buf_get_changedtick(bufnr)
-    ignore_edits = false
-
-    debug({ changedtick_before = changedtick_before, changedtick_after = changedtick_after })
-
-    daemonRevision = daemonRevision + 1
-end
-
 -- Take an operation from the daemon and apply it to the editor.
 local function processOperationForEditor(method, parameters)
     if method == "edit" then
@@ -66,16 +26,15 @@ local function processOperationForEditor(method, parameters)
         local theEditorRevision = parameters.delta.revision
 
         if theEditorRevision == editorRevision then
-            applyDelta(delta)
+            -- Find correct buffer to apply edits to.
+            local bufnr = vim.uri_to_bufnr("file://" .. theFile)
+
+            changetracker.applyDelta(bufnr, delta)
+
+            daemonRevision = daemonRevision + 1
         else
             -- Operation is not up-to-date to our content, skip it!
             -- The daemon will send a transformed one later.
-            print(
-                "Skipping operation, my editor revision is "
-                    .. editorRevision
-                    .. " but operation is for revision "
-                    .. theEditorRevision
-            )
         end
     else
         print("Unknown method: " .. method)
@@ -172,93 +131,25 @@ function EthersyncOpenBuffer()
         vim.bo.eol = false
     end
 
-    prev_lines = vim.api.nvim_buf_get_lines(0, 0, -1, true)
-
     openCurrentBuffer()
 
-    vim.api.nvim_buf_attach(0, false, {
-        on_lines = function(
-            _the_literal_string_lines --[[@diagnostic disable-line]],
-            _buffer_handle --[[@diagnostic disable-line]],
-            _changedtick, --[[@diagnostic disable-line]]
-            first_line,
-            last_line,
-            new_last_line
-        )
-            -- Line counts that we get called with are zero-based.
-            -- last_line and new_last_line are exclusive
+    changetracker.trackChanges(0, function(delta)
+        editorRevision = editorRevision + 1
 
-            debug({ first_line = first_line, last_line = last_line, new_last_line = new_last_line })
-            -- TODO: optimize with a cache
-            local curr_lines = vim.api.nvim_buf_get_lines(0, 0, -1, true)
+        local rev_delta = {
+            delta = delta,
+            revision = daemonRevision,
+        }
 
-            -- Are we currently ignoring edits?
-            if ignore_edits then
-                prev_lines = curr_lines
-                return
-            end
+        local uri = "file://" .. vim.api.nvim_buf_get_name(0)
+        local params = { uri = uri, delta = rev_delta }
 
-            editorRevision = editorRevision + 1
-
-            debug({ curr_lines = curr_lines, prev_lines = prev_lines })
-            local diff = sync.compute_diff(prev_lines, curr_lines, first_line, last_line, new_last_line, "utf-32", "\n")
-            -- line/character indices in diff are zero-based.
-            debug({ diff = diff })
-
-            -- Sometimes, Vim deletes full lines by deleting the last line, plus an imaginary newline at the end. For example, to delete the second line, Vim would delete from (line: 1, column: 0) to (line: 2, column 0).
-            -- But, in the case of deleting the last line, what we expect in the rest of Ethersync is to delete the newline *before* the line.
-            -- So let's change the deleted range to (line: 0, column: [last character of the first line]) to (line: 1, column: [last character of the second line]).
-
-            if diff.range["end"].line == #prev_lines then
-                -- Range spans to a line one after the visible buffer lines.
-                if diff.range["start"].line == 0 then
-                    -- The range starts on the first line, so we can't "shift the range backwards".
-                    -- Instead, we just shorten the range by one character.
-                    diff.range["end"].line = diff.range["end"].line - 1
-                    diff.range["end"].character = vim.fn.strchars(prev_lines[#prev_lines])
-                    if string.sub(diff.text, vim.fn.strchars(diff.text)) == "\n" then
-                        -- The replacement ends with a newline.
-                        -- Drop it, because we shortened the range not to include the newline.
-                        diff.text = string.sub(diff.text, 1, -2)
-                    end
-                else
-                    -- The range doesn't start on the first line.
-                    if diff.range["start"].character == 0 then
-                        -- Operation applies to beginning of line, that means it's possible to shift it back.
-                        -- Modify edit, s.t. not the last \n, but the one before is replaced.
-                        diff.range["start"].line = diff.range["start"].line - 1
-                        diff.range["end"].line = diff.range["end"].line - 1
-                        diff.range["start"].character = vim.fn.strchars(prev_lines[diff.range["start"].line + 1], false)
-                        diff.range["end"].character = vim.fn.strchars(prev_lines[diff.range["end"].line + 1], false)
-                    end
-                end
-            end
-
-            local rev_delta = {
-                delta = {
-                    {
-                        range = {
-                            anchor = diff.range.start,
-                            head = diff.range["end"],
-                        },
-                        replacement = diff.text,
-                    },
-                },
-                revision = daemonRevision,
-            }
-
-            local uri = "file://" .. vim.api.nvim_buf_get_name(0)
-            local params = { uri = uri, delta = rev_delta }
-
-            if online then
-                client.notify("edit", params)
-            else
-                table.insert(opQueueForDaemon, { "edit", params })
-            end
-
-            prev_lines = curr_lines
-        end,
-    })
+        if online then
+            client.notify("edit", params)
+        else
+            table.insert(opQueueForDaemon, { "edit", params })
+        end
+    end)
 end
 
 function EthersyncCloseBuffer()
